chore: rename various dirs and organize files better, using our modern patterns [PT-188049278]
