+++ src/3d-models/museo-500-regular.ts
