+++ src/3d-models/sun-earth-line.ts
