+++ src/assets/earth-simple-0.5k.png
