+++ src/components/city-select.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import CITY_DATA from '../city-data';
-import t, { Language } from '../translate';
+import CITY_DATA from '../utils/city-data';
+import t, { Language } from '../translation/translate';
 
 import '../css/city-select.less';
