+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Seasons from './components/seasons';
-import ScriptingAPI from './scripting-api';
-import ParentMessageAPI from './parent-message-api';
+import ScriptingAPI from './utils/scripting-api';
+import ParentMessageAPI from './utils/parent-message-api';
 
 let seasonsComp = ReactDOM.render(<Seasons/>, document.getElementById('app'));
 let scriptingAPI = new ScriptingAPI(seasonsComp);
