+++ src/utils/solar-system-data.ts
